snowflake/requirements.txt
flask
flask-cors
requests
cerebras-cloud-sdk
python-dotenv